refactor: Pull out team class to separate file
